Move models to shared
