+++ src/Submissions/Api/Models/Shared/ReadCollectionStatusModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Biobanks.Submissions.Api.Models.Shared
+{
+    public class ReadCollectionStatusModel : CollectionStatusModel
+    {
+        public int CollectionCount { get; set; }
+    }
+}
