+++ src/Submissions/Api/Models/Shared/ReadCollectionTypeModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Biobanks.Submissions.Api.Models.Shared
+{
+    public class ReadCollectionTypeModel : CollectionTypeModel
+    {
+        public int CollectionCount { get; set; }
+
+    }
+}
