+++ src/Submissions/Api/Models/Shared/ReadConsentRestrictionModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Biobanks.Submissions.Api.Models.Shared
+{
+    public class ReadConsentRestrictionModel : ConsentRestrictionModel
+    {
+        public int CollectionCount { get; set; }
+
+    }
+}
