--- src/Submissions/Api/Models/Submissions/ReadCollectionStatusModel.cs
@@ -1,7 +1,0 @@
-﻿namespace Biobanks.Submissions.Api.Models.Submissions
-{
-    public class ReadCollectionStatusModel : Shared.CollectionStatusModel
-    {
-        public int CollectionCount { get; set; }
-    }
-}
