--- src/Submissions/Api/Models/Submissions/ReadCollectionTypeModel.cs
@@ -1,8 +1,0 @@
-﻿namespace Biobanks.Submissions.Api.Models.Submissions
-{
-    public class ReadCollectionTypeModel : Shared.CollectionTypeModel
-    {
-        public int CollectionCount { get; set; }
-
-    }
-}
