--- src/Submissions/Api/Models/Submissions/ReadConsentRestrictionModel.cs
@@ -1,8 +1,0 @@
-﻿namespace Biobanks.Submissions.Api.Models.Submissions
-{
-    public class ReadConsentRestrictionModel : Shared.ConsentRestrictionModel
-    {
-        public int CollectionCount { get; set; }
-
-    }
-}
